requirements22.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.111.0
fastapi-cli==0.0.7
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
grpcio==1.68.1
grpcio-tools==1.68.1
h11==0.14.0
hikvisionapi==0.3.2
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.5
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
orjson==3.10.12
packaging==24.2
pillow==10.4.0
pkg_resources==0.0.0
protobuf==5.29.2
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.1
torch==2.4.1+cu121
torchvision==0.19.1+cu121
triton==3.0.0
typer==0.15.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
xmltodict==0.14.2
zipp==3.20.2
